requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.3
feedparser==6.0.11
apscheduler==3.10.4
flask==3.0.2
python-dotenv==1.0.1
